requirements.txt
numpy
pycryptodome
ecdsa
pandas
scikit-learn
matplotlib
streamlit
pyyaml
